Wire the pipeline for custom species; fix real-data crashes
Wiring (new-species path):
- candidates/construct/transcripts/convert-to-transcripts/chkgenes CLIs
  resolve datasets via load_dataset (the candidates CLI previously
  crashed with TypeError: Dataset(path) is missing external_data).
- chkgenes on custom datasets: offline validation against the dataset's
  GTF and annotation tables with difflib near-miss suggestions, no
  Ensembl/mygene network dependency.
- New get_transcripts_generic: --longest/--all transcript selection for
  custom datasets (exact-ID passthrough, gene -> isoform expansion,
  longest-by-sequence via the FASTA), plus resolution through
  registered annotation tables (e.g. ortholog symbol -> transcript,
  case-insensitive). get_transcripts dispatches on dataset type;
  canonical falls back to longest with a note.
- Generic candidates path now allows sibling isoforms of the target's
  gene (parity with the reference path). Without this any multi-isoform
  gene fails screening: every probe "off-targets" its own siblings
  (octopus Och.576.10, 10 isoforms: 0 -> 57 probe pairs).
- Generic candidates path enforces the rRNA/tRNA blocklist
  (check_kmers), mirroring the reference path; warns once when the
  dataset has no blocklist.

Defensive probe-name parsing (IDs may contain underscores and dots):
- New shared helpers probe_coord_exprs/probe_identity_exprs in
  sequtils, used by SAMFrame parsing and screen. The generic path
  always names probes {X}_{X}:start-end, so an even-duplicate prefix
  parses unambiguously for any ID (TRINITY_DN123_c0_g1_i1, STRG.1.1);
  reference-style names split at the last underscore as before.

Real-data fixes (surfaced by the Octopus chierchiae run):
- polars strictness: Float32 UDF declarations -> Float64 (python floats
  are f64), and tm_pairwise's int fallback -> 0.0.
- NumPy 2: OverlapWeighted's uint32 * -1 sentinel array overflows; use
  a float64 array with -1.0 sentinel, which also stops silently
  truncating the sqrt-derived priorities.
- construct: click's multiple=True passes a tuple; isinstance(list)
  check missed it, producing filenames like _ol-2('BamHI', 'KpnI').
- filter-genes (promoted from the deferred list; it blocked the live
  flow): match negative overlaps (_ol-2), read the actual globbed file
  instead of reconstructing a name without the enzyme suffix, and pass
  genes with exactly --min-probes (>= not >).

Live validation on the octopus dataset: candidates -> screen ->
construct -> filter-genes for Och.687.1/Och.958.1/Och.576.10 produced
55/69/54 probes; all 188 screened pairs verified programmatically
(target window == transcript at recorded coords, padlock before splint,
pad_start indexes the splint).

25 new tests (164 total).

+++ src/mkprobes/starmap/simple.py
@@ -64,8 +64,8 @@ def process(codebook: dict[str, list[int]], cwd: Path, n: int = 16):
             dfs[df]
             .with_columns(
                 hairpin=pl.max_horizontal(
-                    pl.col("cons_splint").map_elements(HHP, return_dtype=pl.Float32),
-                    pl.col("cons_pad").map_elements(HHP, return_dtype=pl.Float32),
+                    pl.col("cons_splint").map_elements(HHP, return_dtype=pl.Float64),
+                    pl.col("cons_pad").map_elements(HHP, return_dtype=pl.Float64),
                 ),
                 pos_start=pl.col("pos_start").list.get(0),
             )
